Unify emitting code for binary operator expressions
